Concluded date uses utc millis

+++ app/src/main/java/com/example/shoji/dailytask/provider/TaskContract.java
@@ -38,6 +38,6 @@ public class TaskContract {
 
 
 
-    public static int NOT_CONCLUDED = 0;
-    public static int CONCLUDED = 1;
+    public static long NOT_CONCLUDED = 0;
+    public static long CONCLUDED = 1;
 }
